chore(jans-cedarling): rename folder from data to context_data_api
