+++ jans-cedarling/cedarling/src/context_data_api/entry.rs
